% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{applyLloydMax}
\alias{applyLloydMax}
\title{applyLloydMax}
\usage{
applyLloydMax(fX, K)
}
\arguments{
\item{fX}{(function)     Probability density function of the latent variable}

\item{K}{(positive, int) Possible number of responses}
}
\value{
xkEst (list)     Estimated cutpoints

pkEst (list)     Estimated probabilities

rkEst (list)     Estimated representatives

pkMeans (list)   Mean of estimated probabilities

pkMSEs  (list)   Mean squared errors for convergence plots
}
\description{
Implementation of Lloyd-Max algorithm for simulation of manifest variables
from continuous latent variables.
}
