% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{calc_midpoints}
\alias{calc_midpoints}
\title{Calculate Midpoints}
\usage{
calc_midpoints(repr)
}
\arguments{
\item{repr}{The representation points.}
}
\value{
A vector of midpoints.
}
\description{
Calculate the midpoints of the intervals defined by the representation points.
}
