% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{calc_probs}
\alias{calc_probs}
\title{Calculate Probabilities}
\usage{
calc_probs(density_func, endp)
}
\arguments{
\item{density_func}{The probability density function of the continuous random variable.}

\item{endp}{The end points of the intervals.}
}
\value{
A vector of probabilities for the intervals.
}
\description{
Calculate the discrete probabilities for the given cut points.
}
