% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{compute_distortion}
\alias{compute_distortion}
\title{Compute Distortion}
\usage{
compute_distortion(density_func, midp, repr)
}
\arguments{
\item{density_func}{The probability density function of the continuous random variable \code{X}.}

\item{midp}{The midpoints of the intervals.}

\item{repr}{The representation points of the intervals.}
}
\value{
The distortion (mean-squared error).
}
\description{
Compute the distortion (mean-squared error) for the given representation points.
}
