% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{convert_params}
\alias{convert_params}
\title{Convert parameters.}
\usage{
convert_params(cp)
}
\arguments{
\item{cp}{centered parameters c(mu, sd, gamma1)}
}
\value{
direct parameters c(xi, omega, alpha)
}
\description{
Converts from centered parameters to direct parameters appearing in
the skew-normal density.
}
\seealso{
[sn::cp2dp]
}
