% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert correlation matrix to covariance matrix}
\usage{
cor2cov(corr, s)
}
\arguments{
\item{corr}{correlation matrix}

\item{s}{vector of standard deviations}
}
\value{
covariance matrix
}
\description{
Convert correlation matrix to covariance matrix
}
