% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{delta_skew_normal}
\alias{delta_skew_normal}
\title{Delta parameter of a skew normal distribution}
\usage{
delta_skew_normal(alpha)
}
\arguments{
\item{alpha}{determines the shape}
}
\value{
delta of a skew-normal distribution
}
\description{
Delta parameter of a skew normal distribution
}
