% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{deltaSN}
\alias{deltaSN}
\title{Delta parameter of skew-normal distribution}
\usage{
deltaSN(alpha)
}
\arguments{
\item{alpha}{(real) Determines the skewness or asymmetry; e.g.: alpha=0}
}
\value{
(real)      Delta of skew-normal distribution
}
\description{
Delta parameter of skew-normal distribution
}
