% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{density_sn}
\alias{density_sn}
\title{Probability density function of a skew normal distribution}
\usage{
density_sn(x, xi = 0, omega = 1, alpha = 0)
}
\arguments{
\item{x}{variable}

\item{xi}{determines the location}

\item{omega}{determines the scale}

\item{alpha}{determines the shape}
}
\value{
density at x
}
\description{
Probability density function of a skew normal distribution
}
\seealso{
[sn::dsn()]
}
