% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize_density}
\alias{discretize_density}
\title{Discretize Density}
\usage{
discretize_density(density_func, n_levels, eps = 1e-06)
}
\arguments{
\item{density_func}{The probability density function of the continuous random variable.}

\item{n_levels}{The cardinality of the set of all possible outcomes.}

\item{eps}{The convergence threshold for the algorithm.}
}
\value{
A list containing:
\describe{
  \item{prob}{Discrete probability distribution.}
  \item{endp}{Optimal end points of intervals that partition the domain.}
  \item{repr}{Representation points of the intervals.}
  \item{dist}{Distortion measured as the mean-squared error (MSE).}
}
}
\description{
Transforms the density function of a continuous random variable into a
discrete probability distribution with minimal distortion 
using the Lloyd-Max algorithm.
}
