% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_of_parameters.R
\name{draw_contour}
\alias{draw_contour}
\title{Draw a contour of f given trace}
\usage{
draw_contour(f, x_trace, y_trace)
}
\arguments{
\item{f}{function to find roots}

\item{x_trace}{x coordinates}

\item{y_trace}{y coordinates}
}
\description{
Draw a contour of f given trace
}
