% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{dSN}
\alias{dSN}
\title{Probability density function of a skew-normal distribution.}
\usage{
dSN(x, xi, omega, alpha)
}
\arguments{
\item{x}{(real)               Variable}

\item{xi}{(real)              Determines the location or shift; e.g.: xi=0}

\item{omega}{(positive, real) Determines the scale or dispersion; e.g.: omega=1}

\item{alpha}{(real)           Determines the skewness or asymmetry; e.g.: alpha=0}
}
\value{
(real) Density at x
}
\description{
Probability density function of a skew-normal distribution.
}
