% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_of_parameters.R
\name{estimate_mu_sd}
\alias{estimate_mu_sd}
\title{Estimate mean and standard deviation}
\usage{
estimate_mu_sd(pk, K, gamma1 = 0, trace = FALSE)
}
\arguments{
\item{pk}{a list or a table of proportions of responses}

\item{K}{number of response categories}

\item{gamma1}{skewness of latent variable, 0 by default}

\item{trace}{if TRUE, plots contour of the function and trace iteration steps}
}
\value{
estimated mean and standard deviation of the latent variable
}
\description{
Given proportions pk of responses over categories 1, ..., K:
  pk = (number of responses with value k) / (the number of all responses)
the function estimates the parameters of the latent variable assuming that
the latent variable is normal or skew-normal provided the value of `gamma1`.
}
\examples{
pk <- list("1" = 0.313, "2" = 0.579, "3" = 0.105, "4" = 0.003)
estimate_mu_sd(pk = pk, K = 5)
}
