% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate parameters}
\usage{
estimate_parameters(data, n_levels, skew = 0)
}
\arguments{
\item{data}{The survey responses where the columns correspond to individual items.
Apart from this, `data` can be of almost any class such as
"data.frame" "matrix" or "array".}

\item{n_levels}{The number of response categories, a vector or a number.}

\item{skew}{Marginal skewness of latent variables, defaults to 0.}
}
\value{
A table of estimated parameters for each latent variable.
}
\description{
Given existing survey data, this function estimates the parameters of the latent variables.
}
\seealso{
See also [latent2likert::estimate_mean_and_sd()].
}
