% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{find_representatives}
\alias{find_representatives}
\title{Find Representatives}
\usage{
find_representatives(density_func, midp)
}
\arguments{
\item{density_func}{The probability density function of the continuous random variable.}

\item{midp}{The midpoints of the intervals.}
}
\value{
A vector of representation points.
}
\description{
Find the representation points for the intervals defined by the midpoints.
}
