% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_rand_corr_matrix}
\alias{generate_rand_corr_matrix}
\title{Generate a random p x p correlation matrix}
\usage{
generate_rand_corr_matrix(p)
}
\arguments{
\item{p}{the size of the correlation matrix}
}
\value{
a random p x p correlation matrix
}
\description{
Generate a random p x p correlation matrix
}
