% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_random_cp}
\alias{generate_random_cp}
\title{Get random centered parameters `c(mu, sd, gamma1)` for testing}
\usage{
generate_random_cp()
}
\value{
random centered parameters
}
\description{
Get random centered parameters `c(mu, sd, gamma1)` for testing
}
