% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{get_MSE}
\alias{get_MSE}
\title{Calculate the mean squared error}
\usage{
get_MSE(xk, rk, fX)
}
\arguments{
\item{xk}{vector of cut points}

\item{rk}{vector of representatives}

\item{fX}{probability density function of the latent variable}
}
\value{
mean squared error
}
\description{
Calculate the mean squared error
}
