% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{get_new_rk}
\alias{get_new_rk}
\title{Calculate new representatives `rk` from cut points `xk`.}
\usage{
get_new_rk(xk, fX)
}
\arguments{
\item{xk}{cut points}

\item{fX}{probability density function of the latent variable}
}
\value{
vector `rk` of representatives
}
\description{
Calculate new representatives `rk` from cut points `xk`.
}
