% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{get_new_xk}
\alias{get_new_xk}
\title{Calculate new cut points `xk` from representatives `rk`.}
\usage{
get_new_xk(rk)
}
\arguments{
\item{rk}{vector of representatives}
}
\value{
vector `xk` of cut points
}
\description{
Calculate new cut points `xk` from representatives `rk`.
}
