% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{get_pk}
\alias{get_pk}
\title{Calculate probabilities `pk` from cut points `xk`.}
\usage{
get_pk(xk, fX)
}
\arguments{
\item{xk}{cut points}

\item{fX}{probability density function of the latent variable}
}
\value{
vector of probabilities `pk`
}
\description{
Calculate probabilities `pk` from cut points `xk`.
}
