% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{get_pk_mean}
\alias{get_pk_mean}
\title{Calculate the mean of `pk`}
\usage{
get_pk_mean(pk)
}
\arguments{
\item{pk}{vector of probabilities `pk`}
}
\value{
mean of `pk`
}
\description{
Calculate the mean of `pk`
}
