% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_pk_var}
\alias{get_pk_var}
\title{Calculate variance of probabilities `pk`}
\usage{
get_pk_var(pk)
}
\arguments{
\item{pk}{vector of probabilities}
}
\value{
variance of `pk`
}
\description{
Calculate variance of probabilities `pk`
}
