% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_prop_table}
\alias{get_prop_table}
\title{Get a table of proportions across each possible response.}
\usage{
get_prop_table(data, K)
}
\arguments{
\item{data}{a vector or array of responses}

\item{K}{number of response categories}
}
\value{
table of proportions
}
\description{
Get a table of proportions across each possible response.
}
