% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{get_rand_corr_matrix}
\alias{get_rand_corr_matrix}
\title{Generate a random `p x p` correlation matrix.}
\usage{
get_rand_corr_matrix(p)
}
\arguments{
\item{p}{the size of the correlation matrix}
}
\value{
a random `p x p` correlation matrix
}
\description{
Generate a random `p x p` correlation matrix.
}
