% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_random_cp}
\alias{get_random_cp}
\title{Get random centered parameters `c(mu, sd, gamma1)` for testing.}
\usage{
get_random_cp()
}
\value{
random centered parameters
}
\description{
Get random centered parameters `c(mu, sd, gamma1)` for testing.
}
