% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{get_responses}
\alias{get_responses}
\title{Get responses.}
\usage{
get_responses(n = 10, mu = 0, sd = 1, gamma1 = 0, K = 5, R = 0)
}
\arguments{
\item{n}{number of observations}

\item{mu}{means of latent variables, a vector or a number}

\item{sd}{standard deviations of latent variables, a vector or a number}

\item{gamma1}{marginal skewness of latent variables, a vector or a number.
The values must be between -0.95 and 0.95.}

\item{K}{numbers of response categories, a vector or a number}

\item{R}{correlations between latent variables, 0 by default.
The value can be:
    - a number: same correlation is used between all
                 pairs of latent variables;
    - "random": a random correlation matrix is used;
    - a correlation matrix.}
}
\value{
an array of random responses
}
\description{
Generates a sample of random responses based on parameters of latent variables.
}
\examples{
data <- get_responses(n = 5, mu = c(0, 1, 2))
}
\seealso{
See [responsesR::get_univariate_responses()] for univariate case.
}
