% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{get_univariate_responses}
\alias{get_univariate_responses}
\title{Get univariate responses.}
\usage{
get_univariate_responses(n = 10, mu = 0, sd = 1, gamma1 = 0, K = 5)
}
\arguments{
\item{n}{number of observations}

\item{mu}{mean of latent variable}

\item{sd}{standard deviation of latent variable}

\item{gamma1}{skewness of latent variable, must be between -0.95 and 0.95}

\item{K}{number of response categories, default is 5}
}
\value{
a vector of random responses
}
\description{
Generates a sample of random univariate responses.
}
\seealso{
See [responsesR::simulate_responses()] for an explanation of a
discrete random variable from which samples are taken.
}
