% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{getMSE}
\alias{getMSE}
\title{Calculates mean squared error.}
\usage{
getMSE(xk, rk, fX)
}
\arguments{
\item{xk}{(vector)   Cutpoints}

\item{rk}{(vector)   Representatives}

\item{fX}{(function) Probability density function of the latent variable}
}
\value{
mse (real)   Mean squared error
}
\description{
Calculates mean squared error.
}
