% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{getNewRk}
\alias{getNewRk}
\title{Calculates new representatives rk.}
\usage{
getNewRk(xk, fX)
}
\arguments{
\item{xk}{(vector)   Cutpoints}

\item{fX}{(function) Probability density function of the latent variable}
}
\value{
rk (vector)  Representatives
}
\description{
Calculates new representatives rk.
}
