% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{getNewXk}
\alias{getNewXk}
\title{Calculates new cutpoints xk from representatives rk.}
\usage{
getNewXk(rk)
}
\arguments{
\item{rk}{(vector)  Representatives}
}
\value{
xk (vector) Cutpoints
}
\description{
Calculates new cutpoints xk from representatives rk.
}
