% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{getPk}
\alias{getPk}
\title{Calculates probabilities pk from cutpoints xk.}
\usage{
getPk(xk, fX)
}
\arguments{
\item{xk}{(vector)   Cutpoints}

\item{fX}{(function) Probability density function of the latent variable}
}
\value{
pk (vector)  Probabilities
}
\description{
Calculates probabilities pk from cutpoints xk.
}
