% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{getPkVar}
\alias{getPkVar}
\title{Calculates variance of pk}
\usage{
getPkVar(pk)
}
\arguments{
\item{pk}{(vector) Probabilities}
}
\value{
var (real) Variance of pk
}
\description{
Calculates variance of pk
}
