% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{mean_skew_normal}
\alias{mean_skew_normal}
\title{The mean of skew normal distribution}
\usage{
mean_skew_normal(alpha)
}
\arguments{
\item{alpha}{determines the shape}
}
\value{
mean of a skew-normal distribution
}
\description{
The mean of skew normal distribution
}
