% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{meanSN}
\alias{meanSN}
\title{Mean of skew-normal distribution}
\usage{
meanSN(alpha)
}
\arguments{
\item{alpha}{(real) Determines the skewness or asymmetry; e.g.: alpha=0}
}
\value{
(real)      Mean of skew-normal distribution
}
\description{
Mean of skew-normal distribution
}
