% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pad_levels}
\alias{pad_levels}
\title{Pad missing levels with zeros}
\usage{
pad_levels(pk, K)
}
\arguments{
\item{pk}{proportions or probabilities across possible responses}

\item{K}{number of response categories}
}
\value{
table of proportions across all possible responses
}
\description{
Pad missing levels with zeros
}
