% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{percentify}
\alias{percentify}
\title{Return percentage for a given number}
\usage{
percentify(x, digits = 0)
}
\arguments{
\item{x}{number}

\item{digits}{number of digits}
}
\description{
Return percentage for a given number
}
