% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_of_parameters.R
\name{plot_contour}
\alias{plot_contour}
\title{Plot the contour of function `f` along with provided trace}
\usage{
plot_contour(f, x_trace, y_trace)
}
\arguments{
\item{f}{function to find roots}

\item{x_trace}{x coordinates of the trace}

\item{y_trace}{y coordinates of the trace}
}
\description{
Plot the contour of function `f` along with provided trace
}
