% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_likert_transform}
\alias{plot_likert_transform}
\title{Plot Transformation}
\usage{
plot_likert_transform(n_items, n_levels, mean = 0, sd = 1, skew = 0)
}
\arguments{
\item{n_items:}{Integer. The number of Likert scale items (questions).}

\item{n_levels:}{Integer or vector of integers. The number of response 
             categories (points) for each Likert scale item.

Latent Variables Parameters}

\item{mean:}{Numeric or vector of numerics. Means of the latent variables.
Defaults to 0.}

\item{sd:}{Numeric or vector of numerics. Standard deviations of the latent variables.
Defaults to 1.}

\item{skew:}{Numeric or vector of numerics. Marginal skewness of the latent variables.
Defaults to 0.}
}
\description{
Plots the densities of latent variables in the first row 
and transformed discrete probability distributions below.
}
\details{
Arguments
}
\examples{
plot_likert_transform(n_items = 3, n_levels = c(3, 4, 5))
plot_likert_transform(n_items = 3, n_levels = 5, mean=c(0, 1, 2))
plot_likert_transform(n_items = 3, n_levels = 5, sd=c(0.8, 1, 1.2))
plot_likert_transform(n_items = 3, n_levels = 5, skew=c(-0.5, 0, 0.5))
}
