% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rLikert}
\alias{rLikert}
\title{rLikert}
\usage{
rLikert(size, n_levels, n_items, mean = 0, sd = 1, skew = 0, corr = 0)
}
\arguments{
\item{size:}{Integer. The number of observations (responses) to generate.}

\item{n_levels:}{Integer or vector of integers. The number of response categories 
for each Likert scale item. Defaults to 5.}

\item{n_items:}{Integer. The number of Likert scale items (questions).

Latent Variables Parameters}

\item{mean:}{Numeric or vector of numerics. Means of the latent variables. 
Defaults to 0.}

\item{sd:}{Numeric or vector of numerics. Standard deviations of the latent variables. 
Defaults to 1.}

\item{skew:}{Numeric or vector of numerics. Marginal skewness of the latent variables. 
Defaults to 0.}

\item{corr:}{Numeric or correlation matrix. Correlations between the latent variables. 
Can be a single numeric value representing the same correlation for all pairs, 
"random" for a random correlation matrix, or an actual correlation matrix. 
Defaults to 0.}
}
\description{
Generates an array of random responses to Likert-type questions based on 
specified latent variables.
}
\details{
Arguments
}
