% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{run_Lloyd}
\alias{run_Lloyd}
\title{Implementation of Lloyd's algorithm}
\usage{
run_Lloyd(fX, K)
}
\arguments{
\item{fX}{probability density function of the latent variable}

\item{K}{number of response categories}
}
\value{
xk_estimates list of estimated cut points

pk_estimates list of estimated probabilities

rk_estimates list of estimated representatives

pk_means list of means of estimated probabilities

pk_MS_errors list of mean squared errors for convergence plots
}
\description{
Given a probability density function `fX` of a continuous random variable `X`,
the function returns a discrete random variable with `K` possible outcomes
that best approximates `X` in the mean-square or L2 sense by minimizing MSE.
}
