% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{scale_and_shift}
\alias{scale_and_shift}
\title{Scales and shift cut points.}
\usage{
scale_and_shift(x, dp)
}
\arguments{
\item{x}{variable}

\item{dp}{direct parameters `xi`, `omega`, `alpha`}
}
\value{
shifted and scaled `x`
}
\description{
Scales and shift cut points.
}
