% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{scaleShiftSN}
\alias{scaleShiftSN}
\title{Shifts and scales cutpoints of skew-normal distribution}
\usage{
scaleShiftSN(x, omega, xi, alpha)
}
\arguments{
\item{x}{(real)               Variable}

\item{omega}{(positive, real) Determines the scale or dispersion; e.g.: omega=1}

\item{xi}{(real)              Determines the location or shift; e.g.: xi=0}

\item{alpha}{(real)           Determines the skewness or asymmetry; e.g.: alpha=0}
}
\value{
(real) Shifted and scaled x
}
\description{
Shifts and scales cutpoints of skew-normal distribution
}
