% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{simulate_responses}
\alias{simulate_responses}
\title{Simulate responses.}
\usage{
simulate_responses(K, params)
}
\arguments{
\item{K}{number of response categories}

\item{params}{parameters `xi`, `omega` and `alpha` of skew-normal distribution}
}
\value{
lists of probabilities `pk` and cut points `xk`
}
\description{
Returns a description of a discrete random variable with `K` possible
outcomes that best approximates the latent distribution in the mean-square
sense.
}
\examples{
simulate_responses(K = 5, params = c("mu"=0, "sd"=1, "gamma1"=0))
}
