% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{simulateLikert}
\alias{simulateLikert}
\title{simulateLikert}
\usage{
simulateLikert(K, xi, omega, alpha)
}
\arguments{
\item{K}{(positive, int)      Number of possible responses; e.g.: K=5}

\item{xi}{(real)              Determines the location or shift; e.g.: xi=0}

\item{omega}{(positive, real) Determines the scale or dispersion; e.g.: omega=1}

\item{alpha}{(real)           Determines the skewness or asymmetry; e.g.: alpha=0}
}
\value{
pk, rk, xk (list)     Manifest probabilities, representatives, cutpoints
}
\description{
Simulates Likert scale responses using skew-normal distribution and optimal
discretization.
}
