% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{var_skew_normal}
\alias{var_skew_normal}
\title{Variance of a skew normal distribution}
\usage{
var_skew_normal(alpha)
}
\arguments{
\item{alpha}{determines the shape}
}
\value{
variance of a skew-normal distribution
}
\description{
Variance of a skew normal distribution
}
