% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{varSN}
\alias{varSN}
\title{Variance of skew-normal distribution}
\usage{
varSN(alpha)
}
\arguments{
\item{alpha}{(real) Determines the skewness or asymmetry; e.g.: alpha=0}
}
\value{
(real)      Variance of skew-normal distribution
}
\description{
Variance of skew-normal distribution
}
