% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_of_responses.R
\name{cor2cov}
\alias{cor2cov}
\title{Get covariance matrix from a correlation matrix.}
\usage{
cor2cov(R, s)
}
\arguments{
\item{R}{correlation matrix}

\item{s}{standard deviation vector}
}
\value{
covariance matrix
}
\description{
Get covariance matrix from a correlation matrix.
}
