% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_of_parameters.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate parameters}
\usage{
estimate_parameters(data, K, gamma1 = 0)
}
\arguments{
\item{data}{survey responses, where the columns correspond to individual items.
Apart from this, `data` can be of almost any class such as
"data.frame" "matrix" or "array".}

\item{K}{number of response categories, a vector or a number}

\item{gamma1}{marginal skewness of latent variables, 0 by default}
}
\value{
a table of estimated parameters for each latent variable
}
\description{
Given survey responses, this function estimates the parameters of the latent
variables assuming that the latent variables are normally distributed.
}
\examples{
estimate_parameters(data = c(2, 3, 1, 4, 4, 3, 1, 4, 5, 4), K = 5)
}
\seealso{
See [responsesR::estimate_mu_sd()] for univariate case.
}
